refactor: define command line arguments with simple-parsing
It uses argparse under the hood, but they can be defined with
dataclasses and a big advantage compared to something like Hydra, the
comments are automatically used for the help message.
Everything is type checked, so accessing it can be ensured statically.
Similar to Hydra, it can also accept a config (and also potentially save
the config).

This also makes it much easier to compose multiple configs, e.g. all the
hardware configurations (batch_size, num_workers, num_gpus) are defined
in `config.hardware.HardwareConfig` and can simply be reused by adding
it to a field.

Furthermore, the help message are a grouped by subcategories and also
have the category as a prefix, which is also optional as long as there
is no conflict. So overall it just covers exactly what I want most.

Some of the field definitions can be a bit tedious and could potentially
be improved by parsing it from the comment, similar to Rust's
attributes, but that is already as close to what I wanted considering it
is Python.