feat: limit CPUs similar to taskset with `--cpus`
Just like taskset the available CPUs can be limited by specifying which
ones to use in the `--cpus` option, using the same syntax as a list or
range of CPU indices.

Additionally, the number of workers are divide equally across the GPUs
based on the available visible CPUs.

Example:

```sh
--cpus 0,1 # [0, 1]

--cpus 0-3 # [0, 1, 2, 3]

--cpus 1,8,10-12 # [1, 8, 10, 11, 12]

--cpus 0-10:2 # [0, 2, 4, 6, 8, 10]
```