feat: limit GPUs with `--gpus`
The same principle as for the `--cpus` (limiting CPUs) but for GPUs, so
essentially what `CUDA_VISIBLE_DEVICES` would be used for. But here it's
integrated directly, so no need for environment variables or external
launchers. It is simple assigning the device to each rank and that is
used as a default device.

Specifying `--gpus` will overrule `--num-gpus` and just use the
available GPUs.

Example:

```
--gpus 0,1 # [0, 1]

--gpus 0-3 # [0, 1, 2, 3]

--gpus 1,2,6-8 # [1, 2, 6, 7, 8]

--gpus 0-6:2 # [0, 2, 4, 6]
```