perf: enable optional torch.compile during training
torch.compile can be used with the flag `--compile` to speed up the
training by compiling the model ahead of time. The drawback of this, is
that it requires the input to have a static size, otherwise it will be
recompiled for every batch, making it much slower. So to make sure that
the size is constant, the text/target can be padded to a minimum size by
setting `--text-min-length` to a value that is bigger than the longest
target (in tokens).

Example for this dataset:

```
--compile --text-min-length 48
```