feat: save config as YAML
Save the train config as in log/<name>/config.yaml so that is can be
used later to reproduce the experiment.
Note: There is a small issue with this, since `--config_path` only sets
the default values with the ones in the config, but required arguments
do not count, so they still need to be supplied manually.
It's kind of a drawback from also being able to overwrite the values
easily from the command line. Also some values cannot be overwritten,
since there is no negation, e.g. when `--no-cuda` was given, it will be
set to True automatically and there is no option to revert that, which
means you need to manually edit the config.