feat!: enable fp16 and EMA by default
Flipping the two options because they are pretty much always specified,
so it makes sense to enable them by default and have the option to
disable them.
If the GPU supports Tensor cores (compute capability >= 7.0) fp16 used
automatically, but it can be disabled with `--no-fp16`.
The EMA is also just turned on by default, since there is no real
drawback to it and it turned out to generally bring some improvements.
Now the value for the alpha (α) is specified with `--ema-alpha` and it
can be disabled with `--no-ema`.