refactor: put training code into an extensible trainer
This is very similar to libraries such as PyTorch Lightning, but without
all making it fully general, as it's just the same code refactored into
a class with some minor improvement to make it a little neater.
This was also done to make the transition easier when implementing
a custom model, which does not follow the huggingface interface for
things like input/output to the model.

One major difference between this and PyTorch Lightning, is the fact
that the Trainer contains everything that is needed for the
training/validation steps, but it is not contained in the model itself,
e.g. in Lightning the optimiser is defined on the Model itself (due to
the possibility of having multiple optimiser for each model), but if you
want to be able to modify the learning rate, you would have to pass it
somehow to the model and during inference that would be a useless
parameter, hence that should not be part of the model but the trainer
instead.

Also this trainer does not handle multi-processing or data loading or
any such things, these are still kept in the training script.