fix: set padding in collate correctly
The padding token was initially used because the huggingface tokeniser
automatically added the padding when batch encoding is performed, but
since I encode it first and then pad it in the colllate, this won't do
anything. In order to ignore the padding in the loss, it needed to be
set to -100, which is the ignore index of the cross entropy loss, but
that was not handled properly.
Also makes it configurable, which at the moment is just using the
defaults.