perf: mmap data to avoid processing multiple times
With this the data is loaded and preprocessed once instead of every time
on demand, which means that only one process has to do the work and only
once. Afterwards the result is stored on disk and read with a mmap file,
such that it can be retrieved on demand.