chore: disable tokeniser parallelism for workers
Disables the parallelism for the actual training loop where the data
loader creates workers with the tokenisers. The parallelism will be
disabled regardless, so might as well get rid of the warnings.
Also this has essentially no effect if mmap is used, since the
tokenisation was done prior to this and it will be read from the mmaped
file.