perf: zero grads (set to None) at the end of backward
This saves some memory, since the gradients used to sit in memory until
the next backward pass, where they are zeroed out, but since
zero_grad(set_to_none=True) (default) sets them to None, they won't
occupy any memory when the zeroing is done at the end of each backward
pass.
This also means that for the first call, it needs to be zeroed before
accumulating the gradients, which is just done once at the beginning of
the training loop.