fix: avoid occupying memory of unused GPUs
It turns out that `torch.distributed.barrier()` copies the CUDA data to
the GPU matching the current rank, but when rank != device_id it just
occupies unnecessary memory, especially when that GPU is not being
utilised (because of leaving it free for another user).
Strange behaviour, but to fix it, the barrier needs to be told which
devices are being used for the current process.