ci: change to formatting with ruff
Ruff has now a formatter, so black can be replaced by it.
Unfortunately, the isort replacement (sorting imports) that was already
present is not part of the formatter, but the linter with `--fix`,
meaning that the user needs to run both (fix and format) to format them.
But that at least makes it possible to remove isort and black all
together and only use Ruff.
Hopefully, in future updates the formatting will be unified, but since
this is the very first release (it technically is even a beta), there
are some rough edges. I still prefer to adopt it immediately.